refactor: async로 수정
